requirements.txt
APScheduler==3.11.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
decouple==0.0.7
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
